packages/dnsfuzz/pyproject.toml
[project]
name = "dnsfuzz"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Julien OPPLIGER", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "apex",
    "common",
    "dnspython>=2.8.0",
    "open-asset-model>=1.1.1",
    "open-asset-store>=0.0.2",
    "pyrate-limiter>=4.0.2",
    "termcolor>=3.3.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
common = { workspace = true }
apex = { workspace = true }

[project.scripts]
dnsfuzz = "dnsfuzz.__main__:main"
